REVERTME backends/lll: find and strip literals instead of using NIHLIB populus-end marker.
This is committed for posterity and will be reverted.

The approach tries to determine how much of the compiled deployment
bytecode is literals, and strip them off the tail.

This does not work reliably: e.g. a literal may be defined with a
`lit` keyword, but never used in code (deployment or runtime). In
this case, the definition will be present in the code tree, but
the literal's data won't be appended to bytecode. So, data will
be stripped even if it shouldn't be.

Then, if the literal is only used in deployment code, then the
literal data will be appended to the compiled bytecode; if
if it's used only in runtime code, then it will also be appended.
Without parsing the sources, it's impossible to know whether the
data should or shouldn't be stripped.